Fix module build problems by using tsc-bundle
