+++ tsconfig.json
@@ -2,8 +2,8 @@
     "compilerOptions": {
         "removeComments": true,
         "outDir": "./js/",
+        "outFile": "./js/co-design.js",
         "rootDir": "./js_src/",
-        "target": "ES5",
-        "module": "ES2020"
+        "target": "ES5"
     }
 }
